refactor: 의존성 주입 방식 변경 및 예외 처리
- 기존 필드주입 방식에서 생성자 주입 방식으로 변경하였습니다.
