+++ src/main/java/roomescape/dao/ReservationDAO.java
@@ -9,11 +9,13 @@ import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 import roomescape.dto.RequestReservation;
+import roomescape.exception.BadRequestException;
 import roomescape.model.Reservation;
 
 @Repository
 public class ReservationDAO {
-    private JdbcTemplate jdbcTemplate;
+
+    private final JdbcTemplate jdbcTemplate;
     private final SimpleJdbcInsert jdbcInsert;
 
     public ReservationDAO(JdbcTemplate jdbcTemplate, DataSource dataSource) {
@@ -50,7 +52,21 @@ public class ReservationDAO {
         return reservations;
     }
 
+    private void validateRequestReservation(RequestReservation requestReservation) {
+        if (requestReservation.name() == null || requestReservation.name().isEmpty()) {
+            throw new BadRequestException("이름을 작성해주세요");
+        }
+        if (requestReservation.date() == null || requestReservation.date().isEmpty()) {
+            throw new BadRequestException("날짜를 선택해주세요");
+        }
+        if (requestReservation.time() == null || requestReservation.time().isEmpty()) {
+            throw new BadRequestException("시간을 선택해주세요");
+        }
+    }
+
     public Reservation insert(RequestReservation requestReservation) {
+        validateRequestReservation(requestReservation);
+
         SqlParameterSource params = new MapSqlParameterSource()
                 .addValue("name", requestReservation.name())
                 .addValue("date", requestReservation.date())
@@ -61,10 +77,14 @@ public class ReservationDAO {
         return new Reservation(id, requestReservation.name(), requestReservation.date(), requestReservation.time());
     }
 
-    public int delete(Long id) {
+    public void delete(Long id) {
         String sql = "delete from reservation where id = ?";
 
-        return jdbcTemplate.update(sql, Long.valueOf(id));
+        if (findReservationById(id) == null) {
+            throw new BadRequestException("예약 ID가 존재하지 않습니다.");
+        }
+
+        jdbcTemplate.update(sql, Long.valueOf(id));
     }
 
 }
